+++ CHANGELOG.md
@@ -5,6 +5,23 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [7.0.0] - 2020-09-18
+
+### Added
+
+- New ansi_dark and ansi_light themes
+- Added Text.append_tokens for fast appending of string + Style pairs
+- Added Text.remove_suffix
+- Added Text.append_tokens
+
+### Changed
+
+- Text.tabs_to_spaces was renamed to Text.expand_tabs, which works in place rather than returning a new instance
+- Renamed Column.index to Column.\_index
+- Optimized Style.combine and Style.chain
+- Optimized text rendering by fixing internal cache mechanism
+- Optimized hash generation for Styles
+
 ## [6.2.0] - 2020-09-13
 
 ### Added